requirements.txt
pandas>=2.2.0
scikit-learn>=1.5.0
streamlit>=1.44.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
pypdf>=5.0.0
python-docx>=1.1.0
